Allow importing keras_nlp without tensorflow
This change **does not** break the dependency of tensorflow for all text
preprocessing, so most real world workflows will still require a
tensorflow install.

But this would allow running some modeling in Jax with all preprocessing
done with a separately library.

And most important, this will give friendly, actionable install
instructions the first time something is used that requires a
tensorflow install.